+++ client/src/component/Admin/ProductList.jsx
@@ -1,30 +1,40 @@
 import { useEffect } from "react";
+
 import { DataGrid } from "@mui/x-data-grid";
+
 import "./productList.css";
+
 import { useSelector, useDispatch } from "react-redux";
 import {
   clearErrors,
-  getAdminProduct,
+  getAdminProducts,
   deleteProduct,
+  resetDeleteProduct,
 } from "../../features/product/productSlice";
-import { Link, useNavigate } from "react-router-dom";
+
+import { Link } from "react-router-dom";
+
 import { Button, Typography } from "@mui/material";
 import EditIcon from "@mui/icons-material/Edit";
 import DeleteIcon from "@mui/icons-material/Delete";
+
 import SideBar from "./Sidebar";
 import toast from "react-hot-toast";
 
 const ProductList = () => {
   const dispatch = useDispatch();
-  const navigate = useNavigate();
 
-  const { error, products = [] } = useSelector((state) => state.product);
-  const { error: deleteError, isDeleted } = useSelector(
-    (state) => state.product,
-  );
+  const {
+    error,
+    products = [],
+    isDeleted,
+    loading,
+  } = useSelector((state) => state.product);
 
   const deleteProductHandler = (id) => {
-    dispatch(deleteProduct(id));
+    if (window.confirm("Are you sure you want to delete this product?")) {
+      dispatch(deleteProduct(id));
+    }
   };
 
   useEffect(() => {
@@ -33,29 +43,27 @@ const ProductList = () => {
       dispatch(clearErrors());
     }
 
-    if (deleteError) {
-      toast.error(deleteError);
-      dispatch(clearErrors());
-    }
-
     if (isDeleted) {
       toast.success("Product Deleted Successfully");
-      navigate("/admin/dashboard");
+      dispatch(resetDeleteProduct());
     }
 
-    dispatch(getAdminProduct());
-  }, [dispatch, error, deleteError, isDeleted, navigate]);
+    dispatch(getAdminProducts());
+  }, [dispatch, error, isDeleted]);
 
   const columns = [
-    { field: "id", headerName: "Product ID", minWidth: 200, flex: 0.5 },
-
+    {
+      field: "id",
+      headerName: "Product ID",
+      minWidth: 250,
+      flex: 0.5,
+    },
     {
       field: "name",
       headerName: "Name",
-      minWidth: 350,
+      minWidth: 200,
       flex: 1,
     },
-
     {
       field: "stock",
       headerName: "Stock",
@@ -63,7 +71,6 @@ const ProductList = () => {
       minWidth: 150,
       flex: 0.3,
     },
-
     {
       field: "price",
       headerName: "Price",
@@ -71,7 +78,6 @@ const ProductList = () => {
       minWidth: 200,
       flex: 0.5,
     },
-
     {
       field: "actions",
       headerName: "Actions",
@@ -84,7 +90,10 @@ const ProductList = () => {
             <EditIcon />
           </Link>
 
-          <Button onClick={() => deleteProductHandler(params.row.id)}>
+          <Button
+            color="error"
+            onClick={() => deleteProductHandler(params.row.id)}
+          >
             <DeleteIcon />
           </Button>
         </>
@@ -100,36 +109,24 @@ const ProductList = () => {
   }));
 
   return (
-    <>
-      <div className="dashboard">
-        <SideBar />
+    <div className="dashboard">
+      <SideBar />
 
-        <div className="productListContainer">
-          <Typography id="productListHeading">ALL PRODUCTS</Typography>
+      <div className="productListContainer">
+        <Typography id="productListHeading">ALL PRODUCTS</Typography>
 
-          {/* <DataGrid
-            rows={rows}
-            columns={columns}
-            pageSizeOptions={[10]}
-            initialState={{
-              pagination: { paginationModel: { pageSize: 10 } },
-            }}
-            disableRowSelectionOnClick
-            autoHeight
-            className="productListTable"
-          /> */}
-          <DataGrid
-            rows={rows}
-            columns={columns}
-            pageSizeOptions={[10, 25, 50, 100]}
-            initialState={{
-              pagination: { paginationModel: { pageSize: 10 } },
-            }}
-            autoHeight
-          />
-        </div>
+        <DataGrid
+          rows={rows}
+          columns={columns}
+          loading={loading}
+          pageSizeOptions={[10, 25, 50, 100]}
+          initialState={{
+            pagination: { paginationModel: { pageSize: 10 } },
+          }}
+          autoHeight
+        />
       </div>
-    </>
+    </div>
   );
 };
